refactor: split portfolio.tsx and pseudo-etfs.tsx into folder-based sub-components
Extract page-level sub-components into folder-based organization:
- portfolio.tsx → portfolio/ with portfolio-chart.tsx and performers-section.tsx
- pseudo-etfs.tsx → pseudo-etfs/ with create-form.tsx and etf-card.tsx

Closes #365
